refactor: reset request type(dto -> vo) #7
