+++ Trim-Domain/src/main/java/trim/domains/comment/service/QuestionCommentDomainService.java
@@ -2,12 +2,11 @@ package trim.domains.comment.service;
 
 import trim.domains.comment.domain.QuestionComment;
 import trim.domains.member.domain.Member;
-import trim.domains.comment.dto.request.CreateQuestionCommentRequest;
 import trim.domains.question.domain.Question;
 
 
 public interface QuestionCommentDomainService {
-    Long createQuestionComment(Member writer, Question question, CreateQuestionCommentRequest dto);
+    Long createQuestionComment(Member writer, Question question, String content);
 
     void deleteQuestionComment(Member writer, QuestionComment questionComment);
