+++ Trim-Domain/src/main/java/trim/domains/comment/service/QuestionCommentDomainServiceImpl.java
@@ -7,7 +7,6 @@ import trim.domains.comment.validate.QuestionCommentValidator;
 import trim.domains.member.domain.Member;
 import trim.domains.question.domain.Question;
 import trim.domains.comment.domain.QuestionComment;
-import trim.domains.comment.dto.request.CreateQuestionCommentRequest;
 import trim.domains.comment.repository.QuestionCommentRepository;
 
 @DomainService
@@ -19,8 +18,12 @@ public class QuestionCommentDomainServiceImpl implements QuestionCommentDomainSe
     private final QuestionCommentValidator questionCommentValidator;
 
     @Override
-    public Long createQuestionComment(Member writer, Question question, CreateQuestionCommentRequest dto) {
-        QuestionComment questionComment = dto.from(writer, question);
+    public Long createQuestionComment(Member writer, Question question, String content) {
+        QuestionComment questionComment = QuestionComment.builder()
+                .question(question)
+                .writer(writer)
+                .content(content)
+                .build();
         return questionCommentRepository.save(questionComment).getId();
     }
